backend bez security dijela

+++ src/main/java/com/example/PraksaDrustvenaMreza/controller/CommentController.java
@@ -1,5 +1,6 @@
 package com.example.PraksaDrustvenaMreza.controller;
 
+import com.example.PraksaDrustvenaMreza.service.impl.UserService;
 import org.springframework.http.*;
 import com.example.PraksaDrustvenaMreza.dtos.*;
 import org.springframework.web.bind.annotation.*;
@@ -13,20 +14,13 @@ public class CommentController {
     @Autowired
     private CommentServiceInterface commentService;
 
-    @GetMapping(value = "{id}")
-    public ResponseEntity<CommentDTO> getOneComment(@PathVariable("id") Long id) {
-        try {
-            CommentDTO c = commentService.getOne(id);
-            return new ResponseEntity<>(c, HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-        }
-    }
+    @Autowired
+    private UserService userService;
 
     @PostMapping
     public ResponseEntity<CommentDTO> addComment(@RequestBody AddCommentDTO addCommentDTO) {
         try {
-            CommentDTO c = commentService.save(addCommentDTO);
+            CommentDTO c = commentService.save(addCommentDTO, "brboric99");
             return new ResponseEntity<>(c, HttpStatus.CREATED);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -35,9 +29,15 @@ public class CommentController {
 
     @PatchMapping(value = "{id}")
     public ResponseEntity<CommentDTO> updateComments(@PathVariable("id") Long id, @RequestBody CommentDTO commentDTO){
+        CommentDTO co = commentService.getOne(id);
+        UserDTO u = userService.getOneByUserName("brboric99");
         try{
-            CommentDTO c = commentService.update(commentDTO, id);
-            return new ResponseEntity<>(c, HttpStatus.OK);
+            if(u.getUserName().equals(co.getUserDTO().getUserName())) {
+                CommentDTO c = commentService.update(commentDTO, id);
+                return new ResponseEntity<>(c, HttpStatus.OK);
+            }else{
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -45,9 +45,15 @@ public class CommentController {
 
     @DeleteMapping(value = "{id}")
     public ResponseEntity<Void> deleteComment(@PathVariable("id") Long id){
+        CommentDTO c = commentService.getOne(id);
+        UserDTO u = userService.getOneByUserName("brboric99");
         try{
-            commentService.delete(id);
-            return new ResponseEntity<>(HttpStatus.OK);
+            if(u.getUserName().equals(c.getUserDTO().getUserName())) {
+                commentService.delete(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }else{
+                return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+            }
         }catch (Exception e){
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
